feat: Add equippable slot attribute to items in NexoAddon (dev)
This commit introduces the feature of defining the slot where the item can be equipped.
A new class named 'Equippable' has been added which holds the slot information.
The 'Components' class now includes 'Equippable' instead of a boolean flag, allowing us to store the 'slot' attribute.
The 'slot' attribute is now being fetched from the configuration file in 'ItemConfigUtil'.
The 'PlayerInteractListener' now equips the item to the slot defined in the 'Equippable' object.
This change enhances the flexibility of the system by allowing items to be equippable on different slots.