feat(populators): add support for custom block and tree populators
This commit introduces the ability to generate custom blocks and trees in the world, providing a more immersive and unique experience. The custom blocks and trees are defined in YML configuration files, allowing server administrators to easily customize the generation parameters. These include the maximum and minimum Y coordinates, chance of generation, which blocks can be replaced, and more.

In addition, this commit updates the Nexo dependency version and removes unnecessary files related to the old ore populators.

Lastly, a utility class was created to handle the loading of populator configurations from YML files.

This is a significant enhancement to the existing features of the server, offering a greater degree of customization and control over the world generation process.