feat(populators): add support for furniture in Ore class and CustomOrePopulator
The Ore class now supports both CustomBlockMechanic and FurnitureMechanic.
A new constructor has been added to the Ore class to support FurnitureMechanic.
Also, the CustomOrePopulator has been updated to handle the placement of furniture and custom blocks.
A new enum, CustomType, has been introduced to distinguish between furniture and custom blocks.
The PopulatorsConfigUtil has been updated to handle the creation of Ore objects for both furniture and custom blocks.

This change will allow the generation of both custom blocks and furniture in the world.
The placement of furniture has been handled using Bukkit's scheduler to ensure it is done on the main thread.
A check has been added before placing blocks or furniture to ensure that the block data or furniture is not null.

This update will allow more flexibility in world generation, enriching the player's experience.
PopulatorsConfigUtil has been updated to ensure that only valid custom blocks or furniture are used to create Ore objects.