feat: Support vanilla materials in custom ore generator
Expanded the CustomOrePopulator and PopulatorsConfigUtil to handle vanilla Minecraft materials in addition to custom blocks and furniture. This allows vanilla materials to be specified in the block_populator.yml file and be used by the ore generation system.

In PopulatorsConfigUtil, a check for matching materials has been added, and a new Ore object is now created for matching materials. The Ore class has been extended to include a property for vanilla materials.

In CustomOrePopulator, the placeBlock method has been updated to also place blocks when the vanilla material property of Ore is set.

This change enhances the flexibility of the custom ore generation system and enables the use of vanilla materials in custom ore deposits.