feat: Add id property to CustomTree class
This commit introduces an 'id' field to the CustomTree class.
The constructor also has been modified to include the 'id' parameter.
This change allows each CustomTree object to have a unique identifier,
which can facilitate easier tracking and manipulation of these objects.
The PopulatorsConfigUtil class has been updated accordingly to accommodate
this new parameter in the object creation process.