feat(JukeboxPlayableListener): Improve user interaction with Jukebox
This commit removes an unnecessary import from the file, and modifies the
interaction behaviour with the Jukebox in the game.

Additionally, the action bar now displays a message about the currently
playing track if the item used has lore associated with it.
This should enhance the user experience by providing more in-game information
and preserving the player's inventory.