feat: Add support for tall stringblocks placement
This commit adds the support for the placement of tall stringblocks in the
OrePopulator class. Now, when an ore is defined as tall, it will be placed
one block higher with a TRIPWIRE block underneath. This was achieved by
modifying the placeBlock method and adding a check for the tall property of
the ore.

The tall property was added to the Ore class and is now set in the
PopulatorsConfigUtil class when an ore is created.

This commit also fixes an issue in the BreakDoubleHit class where the hitbox
wasn't being refreshed when a player tries to break a furniture.

The version of the project was updated to 1.2.10 in the pom.xml file.

Minor changes were also made to the NexoAddonCommand class to remove
unnecessary imports.