feat(CustomOrePopulator): improve block placement for tall ores
This commit improves the way blocks are placed for tall ores in the
CustomOrePopulator class. Previously, the block data was simply set
without checking if the location above was air. Now, we first check
if the location above is air before setting the block. This will prevent
the placement of blocks in inappropriate locations.