feat: implement vein generation for ore deposits
In this commit, we've introduced a vein generation feature for the ore deposits in the game. It allows the ores to generate in a clustered way, creating a more realistic in-game mining experience. This is achieved by adding two new fields: veinSize and clusterChance, that can be customized in the configuration file 'block_populator.yml'. veinSize determines the number of blocks in a vein, and clusterChance indicates the chance for cluster placement.

Additionally, we added support for shears tool count as a silk touch enchantment. This feature can be enabled or disabled through the configuration file.

We also added support for jukeboxes. When a jukebox block is broken, it stops playing the sound, and the record inside is dropped. The location of jukeboxes playing music is also stored to ensure the sound stops playing when the block is broken.