feat: add fertilizer component and interaction event
This commit introduces a new fertilizer component that can be added to items.
The fertilizer can increase the growth speed of specified furniture items.

A new NexoFurnitureInteractListener has been added to handle the interaction
of a player with furniture. Specifically, it handles the fertilize action.

The Components class has been extended to hold the new fertilizer component.
The ItemConfigUtil class has been updated to read the new component's
properties from the item's configuration.

The version of the project has been increased to 1.2.15 in the pom.xml file.

Removed swords from the list of tools in BreakWithToolsListener as they
are no longer considered tools for this context.

A minor change was also made to the EquippableListener class to remove
the method getAllHelmets, as it is no longer needed.