feat: remove version specific component support and enhance equipping system
This commit removes the necessity to check for a specific component support
version. This is achieved by moving the version check to individual components
like the JukeboxPlayableListener. Thus allowing all components to load
regardless of the version.
The equipping system has been enhanced by checking and preventing the
interaction with a Nexo block while holding a helmet. The system also now
checks if the player interaction is valid before equipping an item.
In the Fertilize event, a Villager Happy particle effect is added when
fertilizing an item. This gives better feedback to the user for a successful
action. Also, it now returns early if components map or specific item component
is null.
Overall, this commit improves the system's robustness and usability.